+++ ChangeLog.md
@@ -1,3 +1,4 @@
 2019-11
 
+- include coords-table - preliminary
 - initial commit